requirements.txt
## Serialisation / deserialisation

pyserde==0.4.0

## GCP

google-cloud-container==2.4.1
google-cloud-secret-manager==2.5.0
google_cloud_aiplatform==1.1.1
google_cloud_storage==1.38.0

## Data versioning

#dvc[gs]==2.5.0

## Experiment tracking

sacred==0.8.2
pymongo==3.11.4

## Terminal UI

questionary==1.10.0

## TODO: do we need Dill?
#dill==0.3.4
